backend/pyproject.toml
[project]
name = "bobcfc-backend"
version = "0.1.0"
description = "AI Agent Collaboration Platform Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "python-multipart>=0.0.18",
    "redis[hiredis]>=5.0.0",
    "rocketmq-client-python>=2.0.0",
    "minio>=7.2.0",
    "langchain>=1.0.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "google-generativeai>=0.8.0",
    "websockets>=14.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "authlib>=1.3.0",
]
